Finished CRUD for Restaurant backend routes

+++ app/api/restaurant_routes.py
@@ -3,7 +3,7 @@ from app.models import Restaurant, User
 from ..forms.restaurant_form import RestaurantForm
 from datetime import date
 from ..models.db import db
-from flask_login import current_user
+from flask_login import current_user, login_required
 
 
 restaurant_routes = Blueprint('restaurant', __name__)
@@ -14,7 +14,7 @@ def get_all_restaurants():
     """
     Query for all restaurants and returns them in a list of restaurant dictionaries
     """
-    print("----------------CURRENT USER --------------------", {current_user.id})
+
     restaurants = Restaurant.query.all()
     return {'restaurants': [restaurant.to_dict() for restaurant in restaurants]}
 
@@ -26,18 +26,34 @@ def get_restaurant_by_id(id):
     one_restaurant = Restaurant.query.get(id)
 
     if not one_restaurant:
-        return { "message": "Restaurant not found!!!" }
+        return { "message": "Restaurant not found!" }, 404
 
     return one_restaurant.to_dict()
 
 
+@restaurant_routes.route('/current')
+@login_required
+def get_owned_restaurants():
+
+    all_restaurants = Restaurant.query.all()
+    owned_restaurants = [ restaurant.to_dict() for restaurant in all_restaurants if restaurant.owner_id == current_user.id ]
+
+    return { "restaurants": owned_restaurants }
+
+
+
 @restaurant_routes.route('/', methods=["POST"])
+@login_required
 def create_restaurant():
     """Route to post a new restaurant"""
+
+    if current_user is None:
+        return { "message": "Authentication required" }
+
     form = RestaurantForm()
-    # print("----------------CURRENT USER --------------------", current_user)
 
     form["csrf_token"].data = request.cookies["csrf_token"]
+
 
     if form.validate_on_submit():
 
@@ -57,21 +73,53 @@ def create_restaurant():
         )
         db.session.add(new_restaurant)
         db.session.commit()
-        return { "resPost": new_restaurant.to_dict() }
+        return new_restaurant.to_dict(), 201
 
     else:
         print(form.errors)
         return { "errors": form.errors }
 
 
+@restaurant_routes.route("/update/<int:restaurantId>", methods=["GET", "POST"])
+@login_required
+def update_restaurant(restaurantId):
+    form = RestaurantForm()
+
+    form["csrf_token"].data = request.cookies["csrf_token"]
+
+    restaurant_to_update = Restaurant.query.get(restaurantId)
+    if restaurant_to_update.owner_id == current_user.id:
+        if form.validate_on_submit():
+            restaurant_to_update.address = form.data["address"]
+            restaurant_to_update.city = form.data["city"]
+            restaurant_to_update.state = form.data["state"]
+            restaurant_to_update.name = form.data["name"]
+            restaurant_to_update.type = form.data["type"]
+            restaurant_to_update.price = form.data["price"]
+            restaurant_to_update.open_hours = form.data["open_hours"]
+            restaurant_to_update.close_hours = form.data["close_hours"]
+            restaurant_to_update.image_url = form.data["image_url"]
+            db.session.commit()
+            return restaurant_to_update.to_dict()
+
+        else:
+            return { "errors": form.errors }, 400
+
+    else:
+        return { "message": "FORBIDDEN" }, 403
+
+
 @restaurant_routes.route("/delete/<int:restaurantId>")
+@login_required
 def delete(restaurantId):
     restaurant_to_delete = Restaurant.query.get(restaurantId)
 
     if restaurant_to_delete:
-        db.session.delete(restaurant_to_delete)
-        db.session.commit()
-        return { "message": "Delete successful!" }
-
+        if restaurant_to_delete.owner_id == current_user.id:
+            db.session.delete(restaurant_to_delete)
+            db.session.commit()
+            return { "message": "Delete successful!" }
+        else:
+            return { "message": "FORBIDDEN" }, 403
     else:
-        return { "message": "Restaurant not found!" }
+        return { "message": "Restaurant not found!" }, 404
